requirements.txt
asgiref==3.5.2
colorama==0.4.5
dill==0.3.5.1
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.5
typing-extensions==4.3.0
wrapt==1.14.1
